Removed defined call used in conjunction with ensure_packages
The defined function shouldn't be used here because a) order cannot
be guaranteed and b) ensure_packages does the heavy lifting of
making certain only one resource exists in the catalog.

Additionally, this is being called with an array of strings, which
fails for the defined function.  See
https://puppet.com/docs/puppet/5.4/function.html#defined for the
function signature.